+++ booquest-api/src/main/java/com/booquest/booquest_api/application/service/sidejob/GenerateSideJobService.java
@@ -1,7 +1,7 @@
 package com.booquest.booquest_api.application.service.sidejob;
 
-import com.booquest.booquest_api.adapter.in.onboarding.web.sidejob.dto.RegenerateAllSideJobRequest;
-import com.booquest.booquest_api.adapter.in.onboarding.web.sidejob.dto.RegenerateSideJobRequest;
+import com.booquest.booquest_api.adapter.in.sidejob.dto.RegenerateAllSideJobRequest;
+import com.booquest.booquest_api.adapter.in.sidejob.dto.RegenerateSideJobRequest;
 import com.booquest.booquest_api.application.port.in.dto.GenerateSideJobRequest;
 import com.booquest.booquest_api.application.port.in.sidejob.GenerateSideJobUseCase;
 import com.booquest.booquest_api.application.port.in.sidejob.RegenerateSideJobUseCase;
